Refactor comment-pr-deployed.py
